Separate the ArrowRegistry into two classes; it and ArrowStateManager...
The ArrowRegistry maintains compatibility and deprecates its outdated methods. From this point forth, the ArrowStateManager exists to manage instances of AlchemicalArrowEntity, and ArrowRegistry instances of AlchemicalArrow implementations. This is to make a clear distinction between the two functions.

In addition to the above, the ArrowRegistry's former static methods will now be instance methods as there is ultimately no purpose to have these be stateless class members. It's a registry after all.