Use NamespacedKeys to compare arrows, not the ItemStack
This now allows players to rename their arrows, or for server owners to change the lore/name or any facet of the item, without having to worry that all other arrows will be invalidated.

A legacy arrow converter is provided temporarily to ensure any arrows in the world are not lost.